fixed cs61b-sp23 lab03

+++ cs61b_sp23/lab03/src/adventure/SpeciesListStage.java
@@ -98,7 +98,7 @@ public class SpeciesListStage implements AdventureStage {
      */
     public static double arraySimilarity(List<String> listOne, List<String> listTwo) {
         List<String> copy = new ArrayList<>(listOne);
-        int similarObjects = 0;
+        double similarObjects = 0;
         for (String o : listTwo) {
             if (copy.contains(o)) {
                 similarObjects++;
